Update IDC_Consistency_Report.rmd
The summary file was extended with the name of countries. This was tried also in the detailed tables. Not only the loading of the report was in trouble but the filtering was then more complicated as if you select FR for France you get also all countries which have FR in their description (ie CF central A<fr>ican Republic).