Update IDC_Consistency_Report.rmd
I was obliged to add line 143-146 as the output of the rsdmx when calling Eurostat is now different. All the years of the period are in the final output (eventually with NA). This was making lot of undue error when the owner of the country does not have data for some years. Then all the lines were blue...